Move DataPair to graph view.
